730 move unix-specific code to separate mod
